refactor: move parser from context class
